--- data/src/main/java/com/tirexmurina/vkinterntask/data/remote/ProductsResponse.kt
@@ -1,5 +1,0 @@
-package com.tirexmurina.vkinterntask.data.remote
-
-data class ProductsResponse(
-    val products: List<ProductModel>
-)
