+++ data/src/main/java/com/tirexmurina/vkinterntask/data/remote/product/ProductRepositoryImpl.kt
@@ -1,12 +1,11 @@
-package com.tirexmurina.vkinterntask.data.remote
+package com.tirexmurina.vkinterntask.data.remote.product
 
+import com.tirexmurina.vkinterntask.data.remote.DummyjsonAPI
 import com.tirexmurina.vkinterntask.domain.entity.Product
 import com.tirexmurina.vkinterntask.domain.repository.ProductRepository
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.withContext
 
 class ProductRepositoryImpl (
-    private val service: ProductAPI,
+    private val service: DummyjsonAPI,
     private val converter: ProductConverter
 ): ProductRepository {
 
@@ -15,6 +14,10 @@ class ProductRepositoryImpl (
 
     override suspend fun getProductById(id: String): Product =
         converter.convert(service.getProductById("products/$id"))
+
+    override suspend fun getProductsByCategory(category: String, parameters: Map<String, String>): List<Product> =
+        service.getProductsByCategory("products/category/$category", parameters).products.map { converter.convert(it) }
+
 
 
 }
