+++ data/src/main/java/com/tirexmurina/vkinterntask/data/remote/product/ProductsResponse.kt
@@ -1,0 +1,7 @@
+package com.tirexmurina.vkinterntask.data.remote.product
+
+import com.tirexmurina.vkinterntask.data.remote.product.ProductModel
+
+data class ProductsResponse(
+    val products: List<ProductModel>
+)
